JAV-508 & SCB-11 move publish interface from core to common
